+++ NEWS.md
@@ -12,6 +12,7 @@
 * Fixed a bug in `spcov_params()` that yielded output with improper names when a named vector was used as an argument.
 * Fixed a bug in `spautor()` that did not properly coerce `M` if given as a matrix (instead of a vector).
 * Fixed a bug in `esv()` that prevented coercion of `POLYGON`geometries to `POINT` geometries if `data` was an `sf` object.
+* Fixed a bug in `esv()` that did not remove `NA` values from the response.
 
 # spmodel 0.3.0